% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{A sample graph dataset from Wikipedia}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{The starting node of an edge.}
  \item{v2}{The ending node of an edge.}
  \item{w}{The weight or cost of the edge.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\usage{
wiki_graph
}
\description{
A dataset representing a graph with 6 nodes and 18 edges, used as an
example in the Wikipedia article for Dijkstra's algorithm.
}
\keyword{datasets}
