% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm to find Greatest Common Divisor (GCD)}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A positive integer.}

\item{b}{A positive integer.}
}
\value{
The greatest common divisor of `a` and `b`.
}
\description{
This function implements the Euclidean algorithm to find the GCD of two numbers.
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
