% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find the greatest common divisor of two numbers}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric scalar.}

\item{b}{A numeric scalar.}
}
\value{
The greatest common divisor of a and b.
}
\description{
This function implements the Euclidean algorithm to find the greatest
common divisor (GCD) of two numeric scalars.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\references{
The description and pseudocode for the algorithm can be found
  at \url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
